+++ projects/demo-auth/src/app/auth/auth-guard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Inject } from '@angular/core';
 import { Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { UserService, IUser } from '@dlr-eoc/user-info';
+import { UserService, IUser } from '@dlr-eoc/ngx-ukis-ui-clarity';
 import { AlertService } from '../components/global-alert/alert.service';
 import { Observable } from 'rxjs';
 import { BasicAuthService } from './basic-auth.service';