+++ projects/demo-auth/src/app/route-components/route-map/route-map.component.ts
@@ -10,7 +10,7 @@ import { MapOlService } from '@dlr-eoc/map-ol';
 import { ClarityIcons, layersIcon, worldIcon} from '@cds/core/icon';
 import { MapOlComponent } from '@dlr-eoc/map-ol';
 import { ClrVerticalNavModule, ClrStandaloneCdkTrapFocus, ClrNavigationModule, ClrIconModule } from '@clr/angular';
-import { LayerControlComponent, BaseLayerControlComponent } from '@dlr-eoc/layer-control';
+import { LayerControlComponent, BaseLayerControlComponent } from '@dlr-eoc/ngx-ukis-ui-clarity';
 ClarityIcons.addIcons(...[layersIcon, worldIcon]);