+++ projects/map-tools/src/public-api.ts
@@ -1,8 +1,0 @@
-/*
- * Public API Surface of map-tools
- */
-
-export * from './lib/mouse-position/mouse-position.component';
-export * from './lib/navigator/map-navigator.component';
-export * from './lib/projection-switch/projection-switch.component';
-