+++ projects/ngx-ukis-ui-clarity/schematics/ng-add/schema.ts
@@ -1,0 +1,11 @@
+export interface UkisNgAddSchema {
+  // The name of the angular project you want to add @dlr-eoc/ngx-ukis-ui-clarity.
+  project?: string;
+  addClr?: boolean;
+  addFiles?: boolean;
+  routing?: boolean;
+
+  updateFiles?: boolean;
+  addMap?: boolean;
+  auth?: boolean;
+}