+++ projects/ngx-ukis-ui-clarity/src/lib/global-alert/global-alert.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ngx-global-alert',
+  imports: [],
+  templateUrl: './global-alert.component.html',
+  styleUrl: './global-alert.component.scss'
+})
+export class GlobalAlertComponent {
+
+}