+++ projects/ngx-ukis-utilities/src/lib/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './array-reverse.pipe';
+export * from './obj-type.pipe';