+++ scripts/library/utils.ts
@@ -78,16 +78,16 @@ export interface IcheckDepsOutput {
   usedDependencies?: string;
 }
 
-type RecursiveMap = Map<string, RecursiveMap>;
+type RecursiveMap = Map<string, RecursiveMap | null>;
 
 /**
  * replaces all dependency versions in your workspaces with the versions set in main package.json dependencies and devDependencies.
  * Workspace paths are a replacement for the default paths from your package workspaces, and another dependency object in the main package can be used using depsKey
  */
 export function setDependencyVersionsInWorkspaces(MAINPACKAGE: IPackageJSON, packageScope: string, workspacesPaths?: string[], depsKey?: string) {
-  const workspaces = workspacesPaths || MAINPACKAGE?.workspaces.map(p => join(CWD, p.replace(/\\/g, '/'), 'package.json'));
+  const workspaces = workspacesPaths || MAINPACKAGE?.workspaces?.map(p => join(CWD, p.replace(/\\/g, '/'), 'package.json'));
   if (workspaces) {
-    let mainDependencies = Object.assign(MAINPACKAGE.devDependencies, MAINPACKAGE.dependencies);
+    let mainDependencies = Object.assign(MAINPACKAGE.devDependencies || {}, MAINPACKAGE.dependencies);
 
     if (depsKey) {
       mainDependencies = MAINPACKAGE[depsKey];
@@ -97,7 +97,7 @@ export function setDependencyVersionsInWorkspaces(MAINPACKAGE: IPackageJSON, pac
       workspaces.forEach(async path => {
         // todo use all deps in man also dev or add a shared deps
         // and replace also in @dlr-eoc/shared-assets wich not in workspacesPaths
-        await updateWorkspace(path, mainDependencies, MAINPACKAGE.version, packageScope);
+        await updateWorkspace(path, mainDependencies, MAINPACKAGE.version as string, packageScope);
       });
 
     } else {
@@ -223,8 +223,8 @@ export function getProjects(angularJson: WorkspaceSchema) {
  */
 export function dependencyGraph(projects: ICustomWorkspaceProject[], packageScope: string, filterPN?: string[] | false) {
   const projectNames = projects.map(p => p.name);
-  const nodesmapGraph: Map<string, RecursiveMap> = new Map();
-  const edges: [string, string][] = [];
+  const nodesmapGraph: RecursiveMap = new Map();
+  const edges: [string, string | undefined][] = [];
   /**
    * Map of depName: projectName
    */
@@ -232,57 +232,49 @@ export function dependencyGraph(projects: ICustomWorkspaceProject[], packageScop
   projects.forEach((project) => {
     if (existsSync(project.packagePath)) {
       const projectPackage: IPackageJSON = require(project.packagePath);
-      const packageProjectName = projectPackage.name.replace(packageScope, '');
-      nodes.push(packageProjectName);
+      const projectPackageName = projectPackage.name.replace(packageScope, '');
+      nodes.push(projectPackageName);
       // also get projectPackage.devDependencies and projectPackage.peerDependencies
 
-      let projectDeps: string[] = null;
-      let projectDevDeps: string[] = null;
-      let projectPeerDeps: string[] = null;
+      let projectDeps: string[] = [];
+      let projectDevDeps: string[] = [];
+      let projectPeerDeps: string[] = [];
       if (projectPackage.dependencies) {
         projectDeps = Object.keys(projectPackage.dependencies).filter((key) => key.indexOf(packageScope) !== -1).map(key => key.replace(packageScope, ''));
-        addDepsOfProject('dependencies', projectDeps, projectNames, packageProjectName, nodesmapGraph, edges);
+        addDepsOfProject('dependencies', projectDeps, projectNames, projectPackageName, nodesmapGraph, edges);
       }
 
       if (projectPackage.devDependencies) {
         projectDevDeps = Object.keys(projectPackage.devDependencies).filter((key) => key.indexOf(packageScope) !== -1).map(key => key.replace(packageScope, ''));
-        addDepsOfProject('devDependencies', projectDevDeps, projectNames, packageProjectName, nodesmapGraph, edges);
+        addDepsOfProject('devDependencies', projectDevDeps, projectNames, projectPackageName, nodesmapGraph, edges);
       }
 
       if (projectPackage.peerDependencies) {
         projectPeerDeps = Object.keys(projectPackage.peerDependencies).filter((key) => key.indexOf(packageScope) !== -1).map(key => key.replace(packageScope, ''));
-        addDepsOfProject('peerDependencies', projectPeerDeps, projectNames, packageProjectName, nodesmapGraph, edges);
+        addDepsOfProject('peerDependencies', projectPeerDeps, projectNames, projectPackageName, nodesmapGraph, edges);
       }
 
       // if project has no dependencies at all
       if (!projectDeps?.length && !projectDevDeps?.length && !projectPeerDeps?.length) {
-        nodesmapGraph.set(packageProjectName, null);
+        nodesmapGraph.set(projectPackageName, null);
       }
     }
   });
-  // fill all null childs not get in the first iteration
+
   traverseUpdate(nodesmapGraph, nodesmapGraph);
 
   if (filterPN) {
-    const filteredGraph: Map<string, RecursiveMap> = new Map();
+    const filteredGraph: RecursiveMap = new Map();
     filterPN.forEach(projectName => {
       if (nodesmapGraph.has(projectName)) {
-        filteredGraph.set(projectName, nodesmapGraph.get(projectName));
+        filteredGraph.set(projectName, nodesmapGraph.get(projectName)!);
       }
     });
 
-    // edge = [projectName, dep->projectName]
-    // check if project name from filter array is edge[0] -> map to edge[1] and remove duplicates
-    const projectsDeps: string[] = edges.filter(e => filterPN.includes(e[0])).map(e => e[1]).filter((item, index, array) => array.indexOf(item) === index);
-
-    // add the dependencies of the project to the filter array so we can then filter all edges edge[0] based on this list
-    projectsDeps.forEach(item => {
-      if (filterPN.indexOf(item) === -1) {
-        filterPN.push(item);
-      }
-    });
-    const filteredEdges: [string, string][] = edges.filter(e => filterPN.includes(e[0]));
+    /** filterPN is then already included in the filteredEdges */
+    const filteredEdges = edgesFromGraph(filterPN, filteredGraph);
     const filteredNodes = nodes.filter(n => filterPN.includes(n));
+
     return { edges: filteredEdges, nodes: filteredNodes, nodesmapGraph: filteredGraph };
   } else {
     return { edges, nodes, nodesmapGraph };
@@ -290,40 +282,87 @@ export function dependencyGraph(projects: ICustomWorkspaceProject[], packageScop
 
 }
 
-function addDepsOfProject(depsType: string, dependencies: string[], projectNames: string[], packageProjectName: string, nodesmapGraph: Map<string, RecursiveMap>, edges: [string, string][]) {
+
+function edgesFromGraph(parent: string | string[], map: RecursiveMap) {
+  const edges: [string, string | undefined][] = [];
+
+  if (!map) return edges;
+
+  // if parent is an array of roots, run recursively for each
+  if (Array.isArray(parent)) {
+    for (const p of parent) {
+      const hasKey = map.get(p);
+      if (hasKey) {
+        edges.push(...edgesFromGraph(p, hasKey));
+      }
+    }
+
+    const uniqueEdges = Array.from(
+      new Map(edges.map(edge => [edge.join('|'), edge])).values()
+    );
+    return uniqueEdges.sort((a, b) => a[0].localeCompare(b[0]));
+  }
+
+  for (const [key, value] of map.entries()) {
+    edges.push([parent, key]);
+    if (value instanceof Map) {
+      edges.push(...edgesFromGraph(key, value));
+    }
+  }
+
+  const uniqueEdges = Array.from(
+    new Map(edges.map(edge => [edge.join('|'), edge])).values()
+  );
+
+  return uniqueEdges.sort((a, b) => a[0].localeCompare(b[0]));
+}
+
+/** 
+ * check if the project has dependencies, if yes add it as node to the graph 
+ * create edges for all its dependencies
+*/
+function addDepsOfProject(depsType: 'dependencies' | 'devDependencies' | 'peerDependencies', dependencies: string[], projectNames: string[], projectPackageName: string, nodesmapGraph: RecursiveMap, edges: [string, string | undefined][]) {
   if (dependencies.length) {
-    const depsObj: Map<string, any> = new Map();
+    const depsObj: RecursiveMap = new Map();
     dependencies.forEach((dep) => {
       if (projectNames.includes(dep)) {
         // graph only for dependencies not like edges for all;
-        if (depsType === 'dependencies') {
+        if (depsType === 'dependencies' || depsType === 'peerDependencies' || depsType === 'devDependencies') {
           if (nodesmapGraph.has(dep)) {
-            depsObj.set(dep, nodesmapGraph.get(dep));
+            depsObj.set(dep, nodesmapGraph.get(dep)!);
           } else {
             depsObj.set(dep, null);
           }
         }
 
-        edges.push([packageProjectName, dep]);
+        edges.push([projectPackageName, dep]);
       }
     });
     if (depsObj.size) {
-      nodesmapGraph.set(packageProjectName, depsObj);
+      // this runs for all depsType, to create a merge of the depsObj if the graph already has the node
+      if (nodesmapGraph.has(projectPackageName)) {
+        const oldMap = nodesmapGraph.get(projectPackageName);
+        const mergeMap: RecursiveMap = new Map([...oldMap!, ...depsObj]);
+        nodesmapGraph.set(projectPackageName, mergeMap);
+      } else {
+        nodesmapGraph.set(projectPackageName, depsObj);
+      }
     }
   }
 }
 
-
-function traverseUpdate(map: RecursiveMap, lookupTree: Map<string, RecursiveMap>) {
-  if (map.size) {
-    for (const [key, value] of map.entries()) {
-      if (!value) {
+/**
+ * fill all null childs if the would have dependencies - not get in the first iteration for the graph
+ * 
+ */
+function traverseUpdate(graph: RecursiveMap, lookupTree: RecursiveMap) {
+  if (graph.size) {
+    for (const [key, value] of graph.entries()) {
+      if (value === null && lookupTree.has(key)) {
         // if value of the key is null and the key is in the tree, replace the map key with the tree key if it is not null
-        if (lookupTree.has(key)) {
-          const treeKey = lookupTree.get(key);
-          if (treeKey) {
-            map.set(key, treeKey);
-          }
+        const treeKey = lookupTree.get(key);
+        if (treeKey) {
+          graph.set(key, treeKey);
         }
       }
       if (value instanceof Map) {
@@ -335,17 +374,11 @@ function traverseUpdate(map: RecursiveMap, lookupTree: Map<string, RecursiveMap>
 
 
 export function getSortedProjects(projects: ICustomWorkspaceProject[], packageScope: string, filterPN?: string[] | false) {
-  const gne = dependencyGraph(projects, packageScope, filterPN);
+  const { edges } = dependencyGraph(projects, packageScope, filterPN);
   /** An array of directed edges describing a graph e.g. edge1   */
-  //const flattdeps = toposort.array(gne.nodes, gne.edges).reverse();
-  const flattdeps = toposort(gne.edges).reverse();
-  let difference = gne.nodes.filter(x => !flattdeps.includes(x));
-  let sortedProjects = flattdeps;
-  if (difference.length) {
-    // if projects do not have dependencies the ara not in flattdeps so we have to collect them from the nodes array
-    sortedProjects = [...difference, ...flattdeps];
-  }
-  return sortedProjects
+  // maybe check other toposort libs https://www.npmjs.com/search?page=0&q=toposort&sortBy=dependent_count
+  const flattdeps = toposort(edges).reverse();
+  return flattdeps;
 }
 
 /**
@@ -431,17 +464,17 @@ export function formatCheckDepsOutput(error: IcheckDepsOutput, showUsed = false)
   projectPath:  ${error.projectPath}
   missingDependencies: ${error.missingDependencies}`;
 
-  if (error.unusedDependencies.length) {
+  if (error.unusedDependencies?.length) {
     str += `
   peerDependencies: ${error.unusedDependencies}`;
   }
 
-  if (error.unusedDevDependencies.length) {
+  if (error.unusedDevDependencies?.length) {
     str += `
   unusedDevDependencies: ${error.unusedDevDependencies}`;
   }
 
-  if (error.invalidFiles.length > 2) {
+  if (error.invalidFiles?.length && error.invalidFiles.length > 2) {
     str += `
   invalidFiles: ${error.invalidFiles}`;
   }
@@ -479,14 +512,14 @@ function checkTransitiveDependencies(depcheckResults: depcheck.Results, packageS
    * get package.json for each dependency and check if it includes one of the dependencies;
    */
   allDependencies.map(key => {
-    if(key === "projects"){
+    if (key === "projects") {
       console.info(`Check imports - maybe it should be ${packageScope} instead of "projects/"`, depcheckResults.using[key]);
     }
     let packagePath = `${key}/package.json`;
     if (key.includes(packageScope)) {
       packagePath = join(CWD, packagePath.replace(packageScope, 'projects/'));
     }
-    const allPackageDeps = [];
+    const allPackageDeps: string[] = [];
     // Resolve package path and then read package.json with FS because some packages do not list ./package.json on there exports so if we require the package.json an error occurres.
     // Package subpath './package.json' is not defined by "exports"
     const resPackagePath = browserifyResolve(packagePath);